+++ src/getPubkey.h
@@ -5,7 +5,7 @@
 #ifndef _GET_PUBKEY_H_
 #define _GET_PUBKEY_H_
 
-uint32_t read_derivation_path(const uint8_t *dataBuffer, size_t size, uint32_t *derivationPath);
+size_t read_derivation_path(const uint8_t *dataBuffer, size_t size, uint32_t *derivationPath);
 void handleGetPubkey(uint8_t p1,
                      uint8_t p2,
                      uint8_t *dataBuffer,
