+++ src/signMessage.h
@@ -10,7 +10,7 @@ extern uint8_t G_numDerivationPaths;
 void handle_sign_message_receive_apdus(uint8_t p1,
                                        uint8_t p2,
                                        const uint8_t *dataBuffer,
-                                       uint16_t dataLength);
+                                       size_t dataLength);
 
 void handle_sign_message_parse_message(volatile unsigned int *tx);
